+++ YouTrack/AgileSetting.php
@@ -8,7 +8,7 @@ namespace YouTrack;
  * @method string getName
  * @method string setName(string $value)
  *
- * @link https://confluence.jetbrains.com/display/YTD65/Get+List+of+Agile+Boards
+ * @link https://www.jetbrains.com/help/youtrack/standalone/2017.1/Get-List-of-Agile-Boards.html
  */
 class AgileSetting extends BaseObject
 {