+++ YouTrack/Build.php
@@ -16,14 +16,11 @@ class Build extends BundleElement
     public function __construct(\SimpleXMLElement $xml = null, Connection $youtrack = null)
     {
         parent::__construct('build', $xml, $youtrack);
-    }
-
-    protected function update($xml)
-    {
-        parent::update($xml);
-        // todo what timezone is this in? it's not documented.
-        $millis = (int)$xml['assembleDate'];
-        $this->assembleDate = new \DateTime('@'.round($millis / 1000));
+        $this->updateDateAttributes(
+            array(
+                'assembleDate',
+            )
+        );
     }
 
     /**